Rename package unit tests directory.
Rename `integration_tests` package to `package_tests` to avoid
confusion, since those tests are unit tests and not integration tests.
