--- processor/error/integration_tests/doc.go
@@ -1,1 +1,0 @@
-package integration_tests
