+++ processor/error/package_tests/doc.go
@@ -1,0 +1,1 @@
+package package_tests
