+++ processor/transaction/package_tests/TestProcessTransactionEmpty.approved.json
