+++ processor/transaction/package_tests/TestProcessTransactionMinimalPayload.approved.json
