+++ processor/transaction/package_tests/TestProcessTransactionMinimalTrace.approved.json
