+++ processor/transaction/package_tests/json_schema_test.go
@@ -1,4 +1,4 @@
-package integration_tests
+package package_tests
 
 import (
 	"testing"
